Reimplement encoder on top of bytes.Buffer
We anyway need to cache in sync.Pool instances of msgpack Encoder
object, and this object is bound to Writer instance (Buffer in our
case). There's no way to reset encoder to use different buffer, and
there's no way to avoid buffer from being recycled if returned, so it's
easier to use bytes.Buffer which has necessary APIs to avoid extra
allocations (see uber-go/zap#691).

The only allocation left in the benchmark is from msgpack's EncodeTime
function, PR is submitted for it: vmihailenco/msgpack#224